+++ src/Validation/ServerSideLayerValidator.php
@@ -108,7 +108,7 @@ class ServerSideLayerValidator implements LayerValidatorInterface {
 		'blendMode' => [
 			'normal', 'multiply', 'screen', 'overlay', 'darken', 'lighten',
 			'color-dodge', 'color-burn', 'hard-light', 'soft-light', 'difference',
-			'exclusion'
+			'exclusion', 'blur'
 		],
 		'arrowhead' => [ 'none', 'arrow', 'circle', 'diamond', 'triangle' ],
 		'arrowStyle' => [ 'single', 'double', 'none' ],